Minor corrections in javadoc to avoid warnings

+++ plugins/io.sarl.eclipse/src/io/sarl/eclipse/launching/runner/SARLLaunchConfigurationDelegate.java
@@ -319,7 +319,7 @@ public class SARLLaunchConfigurationDelegate extends AbstractJavaLaunchConfigura
 		int index;
 		Pair<IRuntimeClasspathEntry, Integer> pair = getJreEntry(entries, bootEntriesPrepend);
 		jreEntry = pair.getKey();
-		index = pair.getValue();
+		index = pair.getValue().intValue();
 		IRuntimeClasspathEntry[] bootEntriesPrep = JavaRuntime
 				.resolveRuntimeClasspath(
 						bootEntriesPrepend
@@ -415,7 +415,9 @@ public class SARLLaunchConfigurationDelegate extends AbstractJavaLaunchConfigura
 	/** Replies the project SRE from the given configuration.
 	 *
 	 * @param configuration - the configuration to read.
+	 * @param verify  - if true verify the SRE validity, do nothing otherwise
 	 * @return the project SRE or <code>null</code>.
+	 * @throws CoreException 
 	 */
 	private ISREInstall getProjectSpecificSRE(ILaunchConfiguration configuration, boolean verify) throws CoreException {
 		IJavaProject jprj = getJavaProject(configuration);
